sqlite,wallet: move tests out of store, add orphaned blocks test
